pyproject.toml
[tool.poetry]
name = "tech-coach-examples"
version = "0.1.0"
description = ""
authors = ["Arne Sørli <81353974+arneso-ssb@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
papermill = ">=2.4.0"
dapla-toolbelt = ">=1.8.2"
pandas = {extras = ["gcp", "performance"], version = ">=2.0.2"}
duckdb = ">=0.9.1"
tomli = ">=2.0.1"
ssb-fagfunksjoner = ">=0.1.0"
polars = ">=0.20.14"
python-dotenv = ">=1.0.1"
google-cloud-secret-manager = ">=2.19.0"
cryptography = ">=42.0.5"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d", "jupyter"], version = ">=23.3.0"}
isort = ">=5.12.0"
jupyter = ">=1.0.0"
jupytext = ">=1.15.2"
nbstripout = ">=0.6.1"
pre-commit = ">=3.3.1"
pre-commit-hooks = ">=4.4.0"
pytest = ">=7.3.1"
ruff = ">=0.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip_gitignore = true
lines_after_imports = 2
# The line below is needed for jupyter notebooks stored as .py in percent format.
# See https://github.com/PyCQA/isort/issues/1338 for details
treat_comments_as_code = ["# %%"]

[tool.jupytext]
formats = "ipynb,auto:percent"
notebook_metadata_filter = "jupytext.text_representation,-jupytext.text_representation.jupytext_version,-widgets,-varInspector"
cell_metadata_filter = "-papermill,tags"

[tool.ruff]
force-exclude = true  # Apply excludes to pre-commit
show-fixes = true
# src = ["src", "tests"]
target-version = "py310"  # Minimum Python version supported
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]
extend-exclude = [
    "__pycache__",
    "old",
    ".ipynb_checkpoints",
    "noxfile.py",
    "docs/conf.py",
]

# Ruff rules may be customized as desired: https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
select = [
    "A",    # prevent using keywords that clobber python builtins
    "B",    # bugbear: security warnings
    "D",    # pydocstyle
    "E",    # pycodestyle
    "F",    # pyflakes
    "ISC",  # implicit string concatenation
    "UP",   # alert you when better syntax is available in your python version
    "RUF",  # the ruff developer's own rules
]
ignore = [
    "ANN101", # Supress missing-type-self.
    "ANN102", # Supress missing-type-cls.
    "ANN202", # Don't requiere return type annotation for private functions.
    "ANN401", # Allow type annotation with type Any.
    "D100",   # Supress undocumented-public-module. Only doc of public api required.
    "E402",   # Supress module-import-not-at-top-of-file, needed in jupyter notebooks.
    "E501",   # Supress line-too-long warnings: trust black's judgement on this one.
    "I001",   # Don't use ruff to sort imports because of jupytext files.
    "T201",   # Allow print statements.
    "ERA001", # Allow commented-out code, needed for jupytext.
    "INP001", # Don't require __init__.py file.
    "PD901",  # Allow df as variable name.
    "B018",   # Like T201, allow expressions printing variables in notebooks.
    "C408",   # Allow unnecessary collection call since can be more readable.
    "D101",   # Don't require docstring.
    "D102",   # Don't require docstring.
    "D107",   # Don't require docstring.
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pydocstyle]
convention = "google"  # You can also use "numpy".

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"**/tests/*" = [
    "ANN001",  # type annotations don't add value for test functions
    "ANN002",  # type annotations don't add value for test functions
    "ANN003",  # type annotations don't add value for test functions
    "ANN201",  # type annotations don't add value for test functions
    "ANN204",  # type annotations don't add value for test functions
    "ANN205",  # type annotations don't add value for test functions
    "ANN206",  # type annotations don't add value for test functions
    "D100",    # docstrings are overkill for test functions
    "D101",
    "D102",
    "D103",
    "S101",    # asserts are encouraged in pytest
]
